+++ app/build.gradle.kts
@@ -55,6 +55,7 @@ android {
 
 dependencies {
 
+    implementation(libs.androidx.room.common)
     val nav_version = "2.7.7"
 
     implementation("androidx.navigation:navigation-compose:$nav_version")
