+++ app/src/main/java/com/example/marvelapp/network/response/ResponseStatus.kt
@@ -1,4 +1,4 @@
-package com.example.marvelapp.network
+package com.example.marvelapp.network.response
 
 enum class ResponseStatus {
     LOADING, ERROR, DONE
