+++ app/src/main/java/com/example/marvelapp/ui/screens/catalog/CatalogVM.kt
@@ -1,39 +1,41 @@
-package com.example.marvelapp.ui.theme.screens.screen_about
+package com.example.marvelapp.ui.screens.catalog
 
 import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.marvelapp.network.ResponseStatus
 import com.example.marvelapp.network.rep.CatalogRep
+import com.example.marvelapp.network.response.ResponseStatus
 import com.example.marvelapp.ui.components.CatalogDataModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class ScreenAboutVM @Inject constructor(
+class CatalogVM @Inject constructor(
     private val CatalogRep: CatalogRep,
 ) : ViewModel() {
-
     private val _status = MutableLiveData<ResponseStatus>()
     val status: LiveData<ResponseStatus> = _status
 
-    private val _CatalogDataModel = MutableLiveData<CatalogDataModel>()
-    val CatalogDataModel: LiveData<CatalogDataModel> = _CatalogDataModel
+    private val _heroesDataModel = MutableLiveData<CatalogDataModel>()
+    val heroesDataModel: LiveData<CatalogDataModel> = _heroesDataModel
 
+    init {
+        getResponseList()
+    }
 
-    fun getCardById(heroId: Int) {
+    fun getResponseList() {
         viewModelScope.launch {
             try {
                 _status.value = ResponseStatus.LOADING
-                val heroData = CatalogRep.getCardById(heroId)
-                _CatalogDataModel.value = heroData
+                val heroData = CatalogRep.getResponseList()
+                _heroesDataModel.value = heroData
                 _status.value = ResponseStatus.DONE
             } catch (e: Exception) {
                 _status.value = ResponseStatus.ERROR
-                Log.e("ScreenAboutVM", "Error", e)
+                Log.e("CatalogVM", "Error", e)
             }
         }
     }
